[tokenizer] Refactor djl_convert python code
