+++ extensions/tokenizers/src/main/python/.gitignore
@@ -1,4 +1,2 @@
 __pycache__
-model/
-tmp/
-models.json
+*.egg-info/
