+++ extensions/tokenizers/src/main/python/djl_converter/fill_mask_converter.py
@@ -14,7 +14,7 @@ import logging
 
 import torch
 
-from huggingface_converter import HuggingfaceConverter
+from djl_converter.huggingface_converter import HuggingfaceConverter
 
 
 class FillMaskConverter(HuggingfaceConverter):
